requirements.txt
flask==2.3.2
pillow==10.0.0
google-cloud-storage==2.9.0
flask-swagger-ui==4.11.1
werkzeug==2.3.4
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scikit-image==0.20.0
google-auth==2.21.0
python-dotenv==1.0.0